+++ CHANGES.md
@@ -1,3 +1,8 @@
+# Version 0.13.4
+
+## Bugs
+
+
 # Version 0.13.3
 ## Bugs
 - Fixed search for terms that inlcude `@` characters